+++ open-hackathon-client/src/client/logging.conf
@@ -32,7 +32,7 @@ args=(sys.stdout,)
 class=logging.handlers.TimedRotatingFileHandler
 level=DEBUG
 formatter=myFormatter
-args=('/var/log/open-hackathon','midnight',1,14)  # define the log file and position, 'w' means override, 'a' attach to old one
+args=('/var/log/open-hackathon/open-hackathon.log','midnight',1,14)  # define the log file and position, 'w' means override, 'a' attach to old one
 
 #define formatter
 [formatter_myFormatter]
