+++ open-hackathon-server/src/hackathon/views/schema/AdminAzureResource.json
@@ -31,7 +31,7 @@
         "properties": {
           "certificate_id": {
             "type": "string",
-            "pattern":"^\d+$",
+            "pattern":"^[0-9]+$",
             "description": "id of certificate in table HackathonAzureKey"
           }
         }
